Working Backend  - some things still to do
Permalink not set when you load with a post.
We will still need to decide on further data for the back end such as:
- CTA Text
- CTA Link
- Any other meta data?

Next part of this development is to create the theme page that will read
this newsletter data.

Lastly we will need to further the plugin to convert the live page into
a zip file with images.  this is probably a whole other plugin though -
perhaps something exists already!